requirements.txt
flask==3.0.0
pillow==10.1.0
qrcode[pil]==7.4.2
python-dotenv==1.0.0
bakong-khqr==1.0.0
gunicorn==21.2.0